Module Imandrakit_log.Log_event

Events.

The log reporter emits events, which can then be written to various outputs.

module Err = Imandrakit_error.Error

type t = {

lvl : Log_level.t;
(*
Log level.
*)

ts : float;
(*
Timestamp, in seconds, since the UNIX epoch.
*)

msg : string;
(*
Log message.
*)

src : string;
(*
Log source.
*)

meta : (string * string) list;
(*
Additional metadata
*)

}

A log event, which we can store, serialize, send elsewhere, etc.

val pp : Ppx_deriving_runtime.Format.formatter -> t -> Ppx_deriving_runtime.unit

val show : t -> Ppx_deriving_runtime.string

val to_serpack : Imandrakit_ser_pack.Ser.state -> t -> Imandrakit_ser_pack.value

val of_serpack : Imandrakit_ser_pack.Deser.state -> Imandrakit_ser_pack.value -> t

val to_twine_ref : (Imandrakit_twine.Encode.t -> t -> Imandrakit_twine.Encode.immediate) ref

val of_twine_ref : (Imandrakit_twine.Decode.t -> int -> t) ref

val to_twine : Imandrakit_twine.Encode.t -> t -> Imandrakit_twine.Encode.immediate

val of_twine : Imandrakit_twine.Decode.t -> int -> t

type json = Yojson.Safe.t

val to_yojson : t -> json

Convert log events to JSON

val json_error : Err.Kind.t

val of_yojson_ : json -> t Err.result

val of_yojson : json -> (t, string) result